IN-350: recactor: logic for post in feed moved to hook
